+++ NEWS.md
@@ -24,6 +24,7 @@ There is a new family of `sheets_sheet_*()` functions that operate on the (work)
      sheet.
   * `sheets_sheet_names()` returns sheet names.
   * `sheets_sheet_add()` adds one or more sheets.
+  * `sheets_sheet_copy()` copies a sheet.
   * `sheets_sheet_delete()` deletes one or more sheets.
   * `sheets_sheet_resize()` changes the number of rows or columns in a sheet.